Add test case to verify zmq reconnect behaviour (#50)
This PR is a partial duplicate of https://github.com/sonic-net/sonic-dash-ha/pull/84
Since crates/swss-common will be moved in this repo.

### why
During investigating issue #75, I created test case to verify zmq behaviour in handling no connection and connection loss. It is found that with the current swss-common zmq implementation with below attributes, we don't need to handle reconnect explicitly.

- PUSH/PULL model
- ZMQ_IMMEDIATE = 0 (default)
- ZMQ_SNDHWM = 10000

### what this PR does

- add unit test for zmq late connect and reconnect for regression
- remove unneeded code for falling back to ProducerStateTable in test environment since zmq client won't fail if zmq server is not connected.
- update test_utils/README.md with missing argument to start redis in test environment